Unify discovery LLM reviewer providers
